12.21 所有时间功能正常
1. 启用0.6s定时器，中断中读时间
2. 主循环遍历两表，增加一次发送机会避免继电器帧和响应帧碰撞
3. 修正星期计算函数参数类型的错误

+++ time.h
@@ -89,7 +89,7 @@ void fix_date(Time *t)
 
 /* 由年月日计算星期几
    返回值0-6代表周一到周日 */
-u8 calc_weekday(u8 y, u8 m, u8 d)
+u8 calc_weekday(int y, u8 m, u8 d)
 {
     u8 w;
